Changed sky sphere initialisation and functioning camera is not also in project.
Noticed that I was calling glusphere every render call. Created a glusphere object once and stored it in a call list which is then used in render to call the stored list. Create once, call many times. A fully functioning camera is now also in the project, it was at last commit I just forgot to mention it.