Added a Ground class and Model class.
Added a ground model which when called renders a plane of cobble stone textured ground. This is facilitated by the Model class that was also added which loads models from .obj files. The previous commit should also have said that a functioning camera is *NOW also in the project, typo.