Added cages and crates and atmosphere.
Added a groupd of iron bar cages rendered using vertex3f(...) to demonstrate transparency although there is still some slight issue but overall I am happy with the effect. Also added an atmosphere that in itself rotates around the main planet, this was simply a case of rendering another slightly larger sphere at the same location as the planet and then applying a transparent atmosphere texture. Also created a group of crates which are rendered this time using a build index. Changed the ground plane to be slightly smaller as I don't want the scene to feel too sparse.