Added a lot to the scene.
Added: Planks of wood lying across the crates which are built using a build index. Created a bunch of gold coins scattered on the top of one of the crates, these are procedurally generated discs. Added a dragon portal which will be used to demonstrate use of the stencil buffer. Placed loaded model braziers either side of the portal. Placed inside the braiziers hot coals, these are low poly procedurally generated spheres. Added a brown dwarf sun for the planets to orbit. This now adds light to the scene in the form of 4 point lights at seperate locations near the suns surface and gives the scene it's overall reddish ambience. The braziers also provide light coming from them, this was acheived with spotlights used as uplighters to provide a nice lighting effect on the portal. Currently working on stencil buffer.