Added shadows and additonal cams.
Added imposter shadows on three of the candles (centre and left side), I had intended to place imposter shadows with the effect that the current ones have on ALL candles but due to how the texture of the altar was being lit the shadows on the right most candles didn't make sense nor did it look good, and so I didn't place them.

Added 2 additional cameras, one is the planet cam which affords zooming in and out, the other is a simple static cam which faces the altar and dragon protal.

All that remains, is to fix the stencil buffer and have something interactive in the scene, I may possibly add an animated light, although I already have enough lights of types and colours and the scope of this has already creeped enough.