Changed where I was setting up the tex env and added interactivity.
Went through all the classes and changed where I was setting up the tex env, dont think it made any difference it was more for consistency as some classes were doing it in render and other in a initTex func. Now it's all done in the same place namely when the tex is loaded. Also added in a spell effect class which creates a spinning spell effect above the altar provding the user is in the correct location and is looking at the center ritual page and then clicks the left mouse button. All that remains is to try and fix the stencil buffer effect.