Fixed stencil buffer issue.
Hmm...maybe not so much fixed but certainly found a workaround. The issue I believe is due to the fact the I'm only rendering a unit radius sphere over my camera and this is causing problems when trying to get the stencil buffer effect to work as by changing my skysphere to simply be a 500 radius sphere that my enitre scene live inside fixed the issue. The stencil was always working correctly it was the method of my skyshpere that was causing it to break. So now that's fixed everything is done, there is some tidying up and class commenting to take of, but it's done.